Arquivo App refatorado
